requirements.txt
azure-ai-documentintelligence>=1.0.0b1
langchain_community
langchain_openai
flask
scikit-learn
openai
azure-storage-blob
python-dotenv
tiktoken
gunicorn
numpy
matplotlib
easyocr
opencv-python
azure-core
